% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian.R
\name{bn.fit.dag}
\alias{bn.fit.dag}
\title{Bayesian Network Analysis}
\usage{
bn.fit.dag(data, dag, method = "bayes")
}
\arguments{
\item{data}{R data frame to fit the Bayesian network}

\item{dag}{DAG generated by}

\item{method}{Method used to analyze. Default is 'bayes'. Alternative is 'mle'}
}
\value{
bn.fit object
}
\description{
This function is an Interface to the fitting function for Bayesian Networks from bnlearn
}
\examples{
load(simpatdag)
bn.prep.dag(simpatdag)
}
