% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian.R
\name{bn.prep.data}
\alias{bn.prep.data}
\title{Preprocess data}
\usage{
bn.prep.data(
  dag,
  data,
  id = "patient_id",
  time_col = "day",
  factorize = FALSE,
  normalize = FALSE
)
}
\arguments{
\item{dag}{result from bn.prep.dag}

\item{data}{R data frame to fit the Bayesian network}

\item{id}{column name of patient identifier (only needed when lag column)}

\item{time_col}{column name of time (integer expected, only needed when lag column)}

\item{factorize}{specify weather all columns should be factors or not}
}
\value{
data frame
}
\description{
This function preprocesses a dataframe to be used for bn.fit
}
\examples{
load(simpatdag)
bn.prep.dag(simpatdag)
}
