% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_estimation.R
\name{calc_beta}
\alias{calc_beta}
\title{Calc beta}
\usage{
calc_beta(data, psi, outcome, exposure, confounder, id)
}
\arguments{
\item{data}{data frame with values}

\item{psi}{number, given psi}

\item{outcome}{string, defines the outcome variable}

\item{exposure}{string, defines the exposure variable}

\item{confounder}{string vector, defines the confounder variables}

\item{id}{string, identify the unique identifier}
}
\description{
This function fits a geeglm to the data for a given psi. it returns a list of values
}
