% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{comparative.plot}
\alias{comparative.plot}
\title{Comparative Plot}
\usage{
comparative.plot(data, exposure, outcome)
}
\arguments{
\item{data}{defines a data frame.}

\item{exposure}{defines the exposure variable}

\item{outcome}{defines the outcome variable}
}
\description{
This function creates a comparative within a tretment variable.
}
\examples{
comparative.plot(data)
}
