% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjtreat.R
\name{est.effect}
\alias{est.effect}
\title{Generate Treatment Effect
Calculates the treatment effect for given values}
\usage{
est.effect(x_i, exposure_j, tau, gamma)
}
\description{
Generate Treatment Effect
Calculates the treatment effect for given values
}
