% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjtreat.R
\name{estimate.gamma.tau}
\alias{estimate.gamma.tau}
\title{Estimate Gamma and Tau}
\usage{
estimate.gamma.tau(
  data,
  outcome,
  exposure,
  variables,
  bound = 10,
  symmetric = TRUE,
  id = "id",
  time_col = "day",
  one.hot = FALSE
)
}
\arguments{
\item{data}{a data frame holding treatment variables and outcome.}

\item{outcome}{defines the outcome column}

\item{exposure}{defines the exposure variable (level)}

\item{bound}{maximum value for gamma and tau}

\item{symmetric}{identifies if wash-in effect is equal to wash-out effect}
}
\value{
a vector with best gamma and tau.
}
\description{
This function analyzes a patient with time varying treatment.
It analyzes the patient with several wash-in and wash-out effects and returns the best gamma and tau. It is implemented for 2 different treatments.
}
\examples{
# Define Variables
outcome <- "Uncertain_Low_Back_Pain"
exposure <- "treatment"
variables <- c("Activity")
id <- "patient_id"
time_col <- "day"

# use the estimate.gamma.tau function to estimate the best gamma tau values
result <- estimate.gamma.tau(data = simpatdat, outcome = outcome, exposure = exposure, variables = variables, bound = 3, symmetric = TRUE, id=id, time_col = time_col)
# fit the adjusted linear model
fit.adj.lm(data = simpatdat, outcome = outcome, exposure = exposure, variables = variables, id = id, time_col = time_col, effects = result$best)
}
