% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_estimation.R
\name{find_optimum_rec}
\alias{find_optimum_rec}
\title{Find Optimum Recursively}
\usage{
find_optimum_rec(
  data,
  outcome,
  exposure,
  confounder,
  id,
  upper_bound_psi,
  upper_bound_beta,
  lower_bound_psi,
  lower_bound_beta,
  number_it,
  max_number_it,
  epsilon
)
}
\description{
This function finds the best psi value recursively
}
