% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjtreat.R
\name{fit.adj.lm}
\alias{fit.adj.lm}
\title{Fit LM with Adjusted Treatment Effect}
\usage{
fit.adj.lm(
  data,
  outcome,
  exposure,
  variables,
  effects,
  id,
  time_col,
  one.hot = FALSE
)
}
\arguments{
\item{data}{Data Frame}
}
\value{
fitted linear model
}
\description{
This function fits a model with a given treatment effect.
}
