% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjtreat.R
\name{gen_treatment_effect}
\alias{gen_treatment_effect}
\title{Generate Treatment Effect}
\usage{
gen_treatment_effect(data, exposure, gamma, tau, id, time_col)
}
\arguments{
\item{gamma}{wash-in effect (default 1)}

\item{tau}{wash-out effect (default 1)}

\item{treatment}{vector of 1 (treated) and 0 (not treated)}
}
\value{
vector with expected treatment effect
}
\description{
This function generates an expected treatment effect for each time point based on gamma and tau.
}
\examples{
# Generate dummy data
treatment <- rep(c(rep(0,7), rep(1,7)),4)
# run example
gamma <- 5
tau <- 7
gen_treatment_effect(treatment, gamma, tau)
}
