% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_estimation.R
\name{nofgest}
\alias{nofgest}
\title{Nofgest}
\usage{
nofgest(
  data,
  outcome,
  exposure,
  confounder,
  id,
  upper_bound_psi,
  lower_bound_psi,
  max_number_it
)
}
\description{
This function use standard IV estimator via additive marginal structural models to analyse a treatment effect
}
