% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simpatdag}
\alias{simpatdag}
\title{Dag}
\format{
A dagitty string with an example dag for simpatdat
\describe{
  \item{Dag}{dag}
}
}
\source{
Example for simpatdag
}
\usage{
simpatdag
}
\description{
This is an example dag for the data.
}
\keyword{datasets}
