% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{wilcox.nofone}
\alias{wilcox.nofone}
\title{Wilcox test}
\usage{
wilcox.nofone(data, exposure, outcome)
}
\arguments{
\item{data}{dataframe with trail data}

\item{exposure}{identifies the column of exposure variable}

\item{outcome}{identifies the column of outccome variable}
}
\description{
This function performs a wilcox test to identify, if there is an effect of the exposure on the outcome.
}
\references{
``
This function uses the method wilcox.test from stats package. Run '?wilcox.test' for more information.
}
