% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian.R
\name{bn.fit.dag}
\alias{bn.fit.dag}
\title{Bayesian Network Analysis}
\usage{
bn.fit.dag(data, dag, method = "bayes")
}
\arguments{
\item{data}{R data frame to fit the Bayesian network}

\item{dag}{DAG generated by}

\item{method}{Method used to analyze. Default is `bayes`. Alternative is `mle`}
}
\value{
bn.fit object
}
\description{
This function is an Interface to the fitting function for Bayesian Networks from bnlearn
}
\examples{
# specify column names
id <- "patient_id"
time_col <- "day"

# Load data
load("data/simpatdag.rda")
load("data/simpatdat.rda")

# Dag preprocessing
bn.dag <- bn.prep.dag(simpatdag)

# Data Preprocessing (Factorization)
simpatdat$Uncertain_Low_Back_Pain <- as.factor(simpatdat$Uncertain_Low_Back_Pain)
simpatdat$Activity <- cut(simpatdat$Activity, 3, labels=c("low Activity", "middle Activity", "high Activity"))
bn.data <- bn.prep.data(bn.dag, simpatdat, id, time_col)
bn.fit.dag(bn.data, bn.dag, method="bayes")
}
