% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian.R
\name{bn.prep.dag}
\alias{bn.prep.dag}
\title{Prepare Dag for BN Methods}
\usage{
bn.prep.dag(dag)
}
\arguments{
\item{dag}{DAG in a DAGitty format}
}
\value{
bnlearn graph based on DAGitty graph
}
\description{
This function reformat a `DAGitty` to `bn.learn` DAG. This is required for further methods.
}
\examples{
bn.prep.dag(simpatdag)
}
