% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian.R
\name{bn.prep.data}
\alias{bn.prep.data}
\title{Preprocess data}
\usage{
bn.prep.data(
  dag,
  data,
  id = "patient_id",
  time_col = "day",
  factorize = FALSE,
  normalize = FALSE
)
}
\arguments{
\item{dag}{result from bn.prep.dag}

\item{data}{R data frame to fit the Bayesian network}

\item{id}{column name of patient identifier (only needed when lag column)}

\item{time_col}{column name of time (integer expected, only needed when lag column)}

\item{factorize}{specify weather all columns should be factors or not}
}
\value{
data frame
}
\description{
This function is preprocessing a dataframe to be used for bn.fit. For that, columns were transformed into factors and lags were created if required.
}
\examples{
# specify column names
id <- "patient_id"
time_col <- "day"

# Load data
load("data/simpatdag.rda")
load("data/simpatdat.rda")

# Dag preprocessing
bn.dag <- bn.prep.dag(simpatdag)

# Data Preprocessing (Factorization)
simpatdat$Uncertain_Low_Back_Pain <- as.factor(simpatdat$Uncertain_Low_Back_Pain)
simpatdat$Activity <- cut(simpatdat$Activity, 3, labels=c("low Activity", "middle Activity", "high Activity"))
bn.data <- bn.prep.data(bn.dag, simpatdat, id, time_col)
}
