% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian.R
\name{bn.results.samples}
\alias{bn.results.samples}
\title{Calculate Results}
\usage{
bn.results.samples(fitted.model, dag, exposure_levels, verbose = TRUE)
}
\arguments{
\item{fitted.model}{bn.fitted model}

\item{dag}{result from bn.prep.dag}

\item{exposure_levels}{levels to test exposure for}

\item{verbose}{Identify if the function should print the results.}
}
\value{
data frame
}
\description{
This function calculates for each exposure the mean outcome on a sample and was implemented to deliver fast results. Nevertheless, the function should not used extensively for giving results as they are not reproducable.
}
\examples{
bn.results.samples(fitted.bn, bn.dag, exposure_levels, FALSE)
}
