% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{comparative.plot}
\alias{comparative.plot}
\title{Comparative Plot}
\usage{
comparative.plot(
  data,
  exposure,
  outcome,
  method = "boxplot",
  adjustment = 3,
  title = "Comparative Plot"
)
}
\arguments{
\item{data}{defines a data frame.}

\item{exposure}{defines the exposure variable}

\item{outcome}{defines the outcome variable}

\item{method}{one of `boxplot` (default) or  `denseplot`}

\item{adjustment}{adjust parameter for ggplot2::geom_density}

\item{title}{title string for plot}
}
\description{
This function creates a comparative within a tretment variable.
}
\examples{
# Define outcome and exposure column
outcome <- "Uncertain_Low_Back_Pain"
exposure <- "treatment"
# Plot outcome among different exposures
comparative.plot(simpatdat, exposure = exposure, outcome = outcome)
}
