% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjtreat.R
\name{estimate_gamma_tau}
\alias{estimate_gamma_tau}
\title{Estimate Gamma and Tau}
\usage{
estimate_gamma_tau(
  patients_data,
  outcome,
  exposure,
  bound = 10,
  symmetric = TRUE
)
}
\arguments{
\item{outcome}{defines the outcome column}

\item{exposure}{defines the exposure variable (level)}

\item{bound}{maximum value for gamma and tau}

\item{symmetric}{identifies if wash-in effect is equal to wash-out effect}

\item{data}{a data frame holding treatment variables and outcome.}
}
\value{
a vector with best gamma and tau.
}
\description{
This function analyzes a patient with time varying treatment.
It analyzes the patient with several wash-in and wash-out effects and returns the best gamma and tau.
}
\examples{
load(simpatdat)
bound <- 5
estimate_gamma_tau(patient, "Uncertain_Low_Back_Pain", exposure = "treatment")
}
