% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_estimation.R
\name{find_optimum_iterate}
\alias{find_optimum_iterate}
\title{Find Optimum Brute Force}
\usage{
find_optimum_iterate(
  data,
  outcome,
  exposure,
  confounder,
  id,
  upper_bound_psi,
  lower_bound_psi,
  steps
)
}
\description{
This function finds the best psi value brute forced
}
