% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjtreat.R
\name{fit.adj.lm}
\alias{fit.adj.lm}
\title{Fit LM with Adjusted Treatment Effect}
\usage{
fit.adj.lm(
  data,
  outcome,
  exposure,
  variables,
  id,
  time_col,
  effects = NA,
  one.hot = FALSE
)
}
\arguments{
\item{data}{Data Frame with patient data}

\item{outcome}{name of the outcome variable column}

\item{exposure}{name of the exposure variable column}

\item{variables}{vector of names of other variables}

\item{effects}{list of effects for}

\item{one.hot}{boolean weather the exporure
variable is on hot encoded or not-}
}
\value{
fitted linear model
}
\description{
This function fits a linear model with a given treatment effect.
For that, it pre processes the data with the given effect and returns the
fitted linear model. An unadjusted linear model can be generated with setting
`effects` to `NA`.
}
