% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_estimation.R
\name{nofgest}
\alias{nofgest}
\title{Nofgest}
\usage{
nofgest(
  data,
  outcome,
  exposure,
  confounder,
  id,
  method = "iterate",
  upper_bound_psi = 2,
  lower_bound_psi = -2,
  max_number_it = NULL,
  steps = 10,
  corstr = "independence",
  verbose = TRUE
)
}
\arguments{
\item{data}{data frame}

\item{outcome}{outcome variable name}

\item{exposure}{exposure variable name}

\item{confounder}{list of confounders}

\item{id}{patient id}

\item{max_number_it}{max number of iterations}

\item{steps}{number of steps if iteration}

\item{corstr}{a character string specifying the correlation structure. The following are permitted: '"independence"', '"exchangeable"', '"ar1"', '"unstructured"' and '"userdefined"'}

\item{verbose}{Boolean, if prelimnary results should be printed}
}
\description{
This function use standard IV estimator via additive marginal structural models to analyse a treatment effect
}
\examples{
load(simpatdat)
nofgest()
}
