% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simpatdag}
\alias{simpatdag}
\title{Dag}
\format{
A dagitty string with an example dag for the `simpatdat` data set.
\describe{
  \item{Dag}{dag string}
}
}
\source{
This dag represents the causal effects in the example data set `simpatdat`
}
\usage{
simpatdag
}
\description{
This is an example dag for the data set `simpatdat`. It contains an _exposure_ variable (`treatment`), the _outcome_ variable (`Uncertain_Low_Bacl_Pain`) and one confounding variable (`Activity`).
}
\keyword{datasets}
